Fix types for irept::set
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@3578 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ src/util/irep.cpp
@@ -308,9 +308,9 @@ Function: irept::set
 
 \*******************************************************************/
 
-void irept::set(const irep_namet &name, const long value)
+void irept::set(const irep_namet &name, const long long value)
 {
-  add(name).id(i2string((int)value));
+  add(name).id(i2string(value));
 }  
 
 /*******************************************************************\
